Update DB URL stored in Conjur for test app
Secretless expects to get the address without a DB name, and the authenticator
expects to get the full connection string including DB name.

This is because the app using the authenticator client gets the connection
string from the env, where it's injected by Summon directly from Conjur. So
what's stored in Conjur has to be the same as what the app client uses to
connect to the DB (which means it needs the protocol, host, port, and db name).

The app using Secretless treats Secretless like it *is* the DB. Secretless as
of 1.1 only supports `host:[port]` in the address field (and NOT the db name
anymore). If the app client formats its request correctly to include the
protocol, the Secretless host/port, and the db name, Secretless will take care
of redirecting the request to the DB host/port and will pass on the rest of the
info.