Update bash script shebangs
At current, the shebangs in this repo generally refer to `/bin/bash`. OSX comes
by default with an OSX version that is very old (3.2) and upgrading to a newer
version runs into issues with the security protections in Catalina.

This repo depends on bash-lib, which requires that you use bash v4+ - so you
need to install another version of bash than what comes on OSX by default in
order to use bash-lib at all. Unfortunately, even if you do that, bc the
scripts in this repo all prescriptively start with the shebang `#!/bin/bash`,
bash-lib will error on running the `start` script bc you end up using the old
version of bash that comes by default with OSX.

This commit updates the scripts in this repo to use the shebang
`#!/usr/bin/env bash` instead, which means you can install an updated version
of bash with homebrew, you don't have to find a messy workaround for replacing
the standard /bin/bash, and when you run the scripts they will use whatever bash
executable appears first in the running user's $PATH variable.