Remove CONJUR_VERSION from configuration
Previously this repo supported demos of authn-k8s on Conjur Enterprise v4, v5,
and DAP v10+. At current, we only test against DAP v10+.

This commit removes all support for specifying CONJUR_VERSION=4 to use this
repo with Conjur Enterprise v4, and makes it clear in the README that this is
not a supported use case. Where CONJUR_VERSION needs to be specified (still
for Secretless and the Conjur client libraries) it is hardcoded to be 5, but
we hope to remove this need soon too.