Split local authenticator and secretless broker
We have an ability to run the demo with local images of the
authenticator client and the secretless broker.

We would like to run the demo with a local authenticator but with
a secretless-broker image from the `cyberark` registry, in order to
trigger this build in the authenticator client build.