Remove CONJUR_APPLIANCE_URL from authenticator image
The authn-k8s client image does not require the appliance URL to be able to
authenticate with Conjur. In this commit, we remove the appliance URLs from the
Authenticator containers.

Note that the app container will still need the appliance URL to authenticate
with Conjur via the default authenticator, once the access token is available
to the application.

Secretless also still needs the appliance URL, since it instantiates the Go
client using the access token it retrieves via Kubernetes authentication.