Remove redundant RoleBinding for conjur-cluster SA
We were creating a RoleBinding for the service-account running in
the pod of the DAP follower for our tests, while it makes no sense to
do so.

This was done because the ClusterRole wasn't created properly in `kubernetes-conjur-deploy`
but as it is fixed there we should remove it from here.