Test oc DeploymentConfigs
This change adds a test for using DeploymentConfigs OpenShift resources
as application identity for Conjur authn-k8s authentication.

This change depends upon a change in a change to the kubernetes-conjur-deploy
scripts:
    cyberark/kubernetes-conjur-deploy#158
This will allow the authn-k8s plugin to compare an application's
DeploymentConfig with DeploymentConfigs that are permitted (via annotations)
in Conjur policy.

Addresses Issue #115