Adds check for platform before checking OPENSHIFT_VERSION
In the stop script, there is a check for the value of the `OPENSHIFT_VERSION`
environment variable setting without first checking that the `PLATFORM`
environment variable is set to "openshift". This can result in a
undefined variable error.

This change adds a check that PLATFORM is set to "openshift" before
checking OPENSHIFT_VERSION value.