Eliminates duplicate CONJUR_OSS_HELM_RELEASE by using HELM_RELEASE
The scripts currently require that the environment variable
CONJUR_OSS_HELM_RELEASE be set. This environment variable is set in our
CI test scripts, but typical users of these scripts would not be aware
that this setting is required.

Fix is to use the HELM_RELEASE environment variable that users are
already aware of in our CI.

Addresses Issue #133